+++ src/problems/problem010.rs
@@ -1,7 +1,7 @@
 use super::utils::erastostene_sieve;
 
 fn solve(n: u32) -> u128 {
-    let primes = erastostene_sieve(n - 1);
+    let primes = erastostene_sieve((n - 1) as u128);
     let result: u128 = primes.iter().map(|&x| x as u128).sum();
     result
 }
