+++ src/problems/tests/problem007.rs
@@ -1,6 +1,6 @@
 #[test]
 fn test_solve() {
-    let result: u32 = super::solve(6);
-    let expected: u32 = 13;
+    let result: u128 = super::solve(6);
+    let expected: u128 = 13;
     assert_eq!(result, expected)
 }
